+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/sync/internal/SyncManager.kt
@@ -96,6 +96,8 @@ internal class SyncManager(
 
     override val syncedEvents: Flow<List<ChatEvent>> = events
 
+    private val mutex = Mutex()
+
     override fun start() {
         logger.d { "[start] no args" }
         val isDisposed = eventsDisposable?.isDisposed ?: true
@@ -196,10 +198,18 @@ internal class SyncManager(
     }
 
     private suspend fun performSync() {
+        logger.d { "[performSync] no args" }
         val cids = logicRegistry.getActiveChannelsLogic().map { it.cid }.ifEmpty {
             logger.w { "[performSync] no active cids found" }
             repos.selectSyncState(currentUserId)?.activeChannelIds ?: emptyList()
         }
+        mutex.withLock {
+            performSync(cids)
+        }
+    }
+
+    @VisibleForTesting
+    internal suspend fun performSync(cids: List<String>) {
         if (cids.isEmpty()) {
             logger.w { "[performSync] rejected (cids is empty)" }
             return
@@ -213,29 +223,32 @@ internal class SyncManager(
         } else {
             chatClient.getSyncHistory(cids, lastSyncAt).await()
         }
-
-        when (result) {
-            is Result.Success -> {
-                val sortedEvents = result.value.sortedBy { it.createdAt }
-                logger.d { "[performSync] succeed. events: ${sortedEvents.size}" }
-                val latestEvent = sortedEvents.lastOrNull()
-                val latestEventDate = latestEvent?.createdAt ?: Date()
-                val rawLatestEventDate = latestEvent?.rawCreatedAt
-                updateLastSyncedDate(latestEventDate, rawLatestEventDate)
-                sortedEvents.forEach {
-                    if (it is MarkAllReadEvent) {
-                        updateAllReadStateForDate(it.user.id, it.createdAt)
-                    }
-                }
-                if (sortedEvents.isNotEmpty() && rawLastSyncAt != null && rawLastSyncAt != rawLatestEventDate) {
-                    events.emit(sortedEvents)
-                    logger.v { "[performSync] events emission completed" }
-                } else {
-                    logger.v { "[performSync] no events to emit" }
-                }
-            }
-            is Result.Failure -> logger.e { "[performSync] failed(${result.value})" }
+        if (result !is Result.Success) {
+            logger.e { "[performSync] failed($result)" }
+            return
         }
+        val sortedEvents = result.value.sortedBy { it.createdAt }
+        logger.d { "[performSync] succeed; events.size: ${sortedEvents.size}" }
+
+        val latestEvent = sortedEvents.lastOrNull()
+        val latestEventDate = latestEvent?.createdAt ?: Date()
+        val rawLatestEventDate = latestEvent?.rawCreatedAt
+        updateLastSyncedDate(latestEventDate, rawLatestEventDate)
+        sortedEvents.forEach {
+            if (it is MarkAllReadEvent) {
+                updateAllReadStateForDate(it.user.id, it.createdAt)
+            }
+        }
+        if (sortedEvents.isEmpty()) {
+            logger.w { "[performSync] rejected (no events to emit)" }
+            return
+        }
+        if (rawLastSyncAt == rawLatestEventDate) {
+            logger.w { "[performSync] rejected (rawLatestEventDate equals to rawLastSyncAt)" }
+            return
+        }
+        events.emit(sortedEvents)
+        logger.v { "[performSync] events emission completed" }
     }
 
     /**